+++ lib/puppet/type/cfsystem_memory_calc.rb
@@ -1,0 +1,26 @@
+Puppet::Type.newtype(:cfsystem_memory_calc) do
+    desc "Calculate memory scopes based on configured weights"
+    
+    autorequire(:cfsystem_flush_config) do
+        ['begin']
+    end
+    autonotify(:cfsystem_flush_config) do
+        ['commit']
+    end    
+    
+    ensurable do
+        defaultvalues
+        defaultto :absent
+    end
+
+    newparam(:name) do
+        desc "Must always be 'total'"
+        isnamevar
+        
+        validate do |value|
+            unless value == 'total'
+                raise ArgumentError, "The only title allowed is 'total'"
+            end
+        end
+    end
+end